requirements/base.txt
Django==4.2
psycopg2==2.9.6
celery==5.3.1         
django-celery-beat==2.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
drf-yasg==1.21.5
setuptools
coreapi==2.3.3
coreschema==0.0.4
redis>=4.5.1
legacy-cgi==2.6.2